requirements.txt
django
django-model-utils==4.0.0
djangorestframework